replace cPickle import with a try... except.. for pickle.
In Python3, the pickle module itself searches if an optimized version
is available in the system. The user does not need to indicate this to
the system, as in Py2 (by importing cPickle instead of pickle).

See: http://docs.pythonsprints.com/python3_porting/py-porting.html

Also, pickling files are in binary format. Current code reads/writes
files with implicitly assuming it's binary. In Py3, this should be
made explicit in the flag (i.e., "open(file, "rb").
See: http://pythonconquerstheuniverse.wordpress.com/2011/09/07/python3-pickling/

Finally, in dose/simulation_calls.py, made explicit that the
inspect.stack() call is to return the basename only. On my system,
it returns the full path, giving an error and no copying of the script
file to the Simulations directory.

With this commit, Example 01 now works under Python 3.4.2 on my system.
(NB not tested for Py2 functioning....)