Replace xrange() with range()
In Python3, the range function has become a generator, no more need
to use xrange() explicitly. See http://python3porting.com/differences.html

For Py3 code, this changes nothing. For Py2 code, however, the generator
is replaced by a list -- and for long lists, this might clog memory and
result in performance issues. As most (all?) xrange iterations appear to
be over small lists, I don't see major issue with this modification for
Py2 users. (Exception is perhaps dose/dose_entities.py, where iteration
over pop_size is now a list. If pop_size is very large, this might cause
an issue #NEEDS VERIFICATION#

Alternatively, the six module can be used to keep xrange in both Py2, Py3
working as it was meant to be. But this requires installation of an extra
module, which should be avoided if possible.