Replace division operator (/) with true division operator (//)
See PEP-238, and here: https://docs.python.org/2/library/stdtypes.html
In this particular case, it seems that when group is odd, the value
of len(group)/2 becomes a float, which throws an error in Py3. With
replacing the / for // operator, the program runs.

I am not sure why we need to divide group by two here: as I understand,
this is to split the population into two groups -- but I cannot deter-
mine if, in case of odd group length, some organisms are not mating here.

Anyway, this fix removes the TypeError and program runs... Modified in
all (example) files where this line occurs.